Rename requires to needs (#82)
* Adding fix for #75, rename requires to needs

* Removing check for install commands (should still be run in subproject)

+++ tests/IniTest.cpp
@@ -248,7 +248,7 @@ TEST_F(TApp, IniGetNoRemaining) {
     int two = 0;
     app.add_option("--two", two);
     EXPECT_NO_THROW(run());
-    EXPECT_EQ(app.remaining().size(), 0);
+    EXPECT_EQ(app.remaining().size(), (size_t)0);
 }
 
 TEST_F(TApp, IniNotRequiredNotDefault) {
